secure_app/requirements.txt
flask>=3.0.0
flask-wtf>=1.2.0
werkzeug>=3.0.0
markupsafe>=2.1.0
python-dotenv>=1.0.0
